move shared blocks functions to BananaUtils file
