+++ CadEditor/settings_nes/banana_prince/Settings_Banana Prince-1-2.cs
@@ -1,5 +1,7 @@
 using CadEditor;
 using System.Collections.Generic;
+//css_include shared_settings/SharedUtils.cs;
+//css_include banana_prince/BananaUtils.cs;
 
 public class Data
 { 
@@ -8,89 +10,21 @@ public class Data
   public bool isBlockEditorEnabled()    { return false; }
   public bool isEnemyEditorEnabled()    { return false; }
   
-  public GetVideoPageAddrFunc getVideoPageAddrFunc() { return getVideoAddress; }
-  public GetVideoChunkFunc    getVideoChunkFunc()    { return getVideoChunk;   }
+  public GetVideoPageAddrFunc getVideoPageAddrFunc() { return SharedUtils.fakeVideoAddr(); }
+  public GetVideoChunkFunc    getVideoChunkFunc()    { return SharedUtils.getVideoChunk(new[] {"chr1.bin"}); }
   public SetVideoChunkFunc    setVideoChunkFunc()    { return null; }
   
   public OffsetRec getBlocksOffset()    { return new OffsetRec(0, 1 , 0x1000);  }
   public int getBlocksCount()           { return 64; }
   
-  public GetBigBlocksFunc     getBigBlocksFunc() { return getBigBlocks;}
-  public SetBigBlocksFunc     setBigBlocksFunc() { return setBigBlocks;}
-  public GetBlocksFunc        getBlocksFunc() { return getBlocksConsts;}
+  public GetBigBlocksFunc     getBigBlocksFunc() { return BananaUtils.getBigBlocks;}
+  public SetBigBlocksFunc     setBigBlocksFunc() { return BananaUtils.setBigBlocks;}
+  public GetBlocksFunc        getBlocksFunc() { return BananaUtils.getBlocksConsts;}
   public SetBlocksFunc        setBlocksFunc() { return null;}
-  public GetPalFunc           getPalFunc()    { return getPallete;}
+  public GetPalFunc           getPalFunc()           { return SharedUtils.readPalFromBin(new[] {"pal1.bin"}); }
   public SetPalFunc           setPalFunc()    { return null;}
   
   public OffsetRec getBigBlocksOffset() { return new OffsetRec(0x463C, 1 , 0x4000); }
   public int getPalBytesAddr()          { return 0x4964; }
   public int getBigBlocksCount() { return 202; }
-  
-  //----------------------------------------------------------------------------
-  public ObjRec[] getBlocksConsts(int blockIndex)
-  {
-        var objects = new ObjRec[getBlocksCount()];
-        for (int i = 0; i < objects.Length; i++)
-        {
-            var indexes  = new int[4];
-            var palBytes = new int[1];
-            int bi = (i/8)*32 + i%8 * 2;
-            indexes[0] = bi;
-            indexes[2] = bi + 1;
-            indexes[1] = bi + 16;
-            indexes[3] = bi + 17;
-            objects[i] = new ObjRec(2, 2, 0, indexes, palBytes);
-        }
-        return objects;
-  }
-  
-  public BigBlock[] getBigBlocks(int bigTileIndex)
-  {
-    var data = Utils.readLinearBigBlockData(0, bigTileIndex);
-    var bb = Utils.unlinearizeBigBlocks<BigBlockWithPal>(data, 2, 2);
-    for (int i = 0; i < bb.Length; i++)
-    {
-      int palByte = Globals.romdata[getPalBytesAddr() + i];
-      bb[i].palBytes[0] = palByte >> 0 & 0x3;
-      bb[i].palBytes[1] = palByte >> 2 & 0x3;
-      bb[i].palBytes[2] = palByte >> 4 & 0x3;
-      bb[i].palBytes[3] = palByte >> 6 & 0x3;
-    }
-    return bb;
-  }
-  
-  public void setBigBlocks(int bigTileIndex, BigBlock[] bigBlockIndexes)
-  {
-    var bigBlocksAddr = ConfigScript.getBigTilesAddr(0, bigTileIndex);
-    var data = Utils.linearizeBigBlocks(bigBlockIndexes);
-    
-    int size = data.Length;
-    int addr = ConfigScript.getBigTilesAddr(0, bigTileIndex);
-    for (int i = 0; i < size; i++)
-    {
-        Globals.romdata[addr + i] =  data[i];
-    }
-    //save pal bytes
-    for (int i = 0; i < bigBlockIndexes.Length; i++)
-    {
-      var bb = bigBlockIndexes[i] as BigBlockWithPal;
-      int palByte = bb.palBytes[0] | bb.palBytes[1] << 2 | bb.palBytes[2]<<4 | bb.palBytes[3]<< 6;
-      Globals.romdata[getPalBytesAddr() + i] = (byte)palByte;
-    }
-  }
-  
-  public byte[] getPallete(int palId)
-  {
-      return Utils.readBinFile("pal1.bin");
-  }
-  
-  public int getVideoAddress(int id)
-  {
-    return -1;
-  }
-  
-  public byte[] getVideoChunk(int videoPageId)
-  {
-     return Utils.readVideoBankFromFile("chr1.bin", videoPageId);
-  }
 }
